backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiohttp==3.9.1
beautifulsoup4==4.12.2
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.5
tiktoken==0.5.2
python-multipart==0.0.6
